refactor: 공통 컴포넌트 및 UI 정리 (#71)
