+++ src/components/common/Card.tsx
@@ -1,5 +1,5 @@
 import type { VariantProps } from 'class-variance-authority'
-import { cn } from '@/lib/utils'
+import { cn } from '@/lib'
 import { cardVariants } from '@/constants'
 
 export interface CardProps
