+++ src/components/common/CheckBox.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Check } from 'lucide-react'
 import type { VariantProps } from 'class-variance-authority'
-import { cn } from '@/lib/utils'
+import { cn } from '@/lib'
 import { checkboxVariants } from '@/constants'
 
 export interface CheckboxProps
