+++ src/components/common/index.ts
@@ -1,11 +1,11 @@
-export * from './Badge.tsx'
-export * from './Breadcrumb.tsx'
-export * from './Button.tsx'
-export * from './CheckBox.tsx'
-export * from './Card.tsx'
-export * from './Dropdown.tsx'
-export * from './IconDropdown.tsx'
-export * from './Input.tsx'
-export * from './Modal.tsx'
-export * from './RangeSlider.tsx'
-export * from './ToastAlert.tsx'
+export * from './Badge'
+export * from './Breadcrumb'
+export * from './Button'
+export * from './CheckBox'
+export * from './Card'
+export * from './Dropdown'
+export * from './IconDropdown'
+export * from './Input'
+export * from './Modal'
+export * from './RangeSlider'
+export * from './ToastAlert'
